requirements.txt
streamlit
flask
spacy
transformers
apscheduler
ortools
email-validator
python-dotenv
requests
uuid
matplotlib
networkx 